Split entity.hpp/.cpp into three
