+++ src/game_api/entities_items.cpp
@@ -4,9 +4,10 @@
 #include <type_traits> // for move
 #include <vector>      // for vector, _Vector_iterator, _Vector_cons...
 
-#include "entities_chars.hpp" // for PowerupCapable
-#include "entity.hpp"         // for HookWithId, EntityHooksInfo, to_id, SHAPE
-#include "vtable_hook.hpp"    // for hook_vtable
+#include "entities_chars.hpp"    // for PowerupCapable
+#include "entity.hpp"            // for to_id, SHAPE
+#include "entity_hooks_info.hpp" // for HookWithId, EntityHooksInfo
+#include "vtable_hook.hpp"       // for hook_vtable
 
 void ParachutePowerup::deploy()
 {
