+++ src/game_api/state.cpp
@@ -9,22 +9,23 @@
 #include <string>      // for allocator, operator""sv, operator""s
 #include <type_traits> // for move
 
-#include "entities_chars.hpp"   // for Player
-#include "entity.hpp"           // for to_id, Entity, HookWithId, EntityDB
-#include "game_manager.hpp"     // for get_game_manager, GameManager, SaveR...
-#include "items.hpp"            // for Items, SelectPlayerSlot
-#include "level_api.hpp"        // for LevelGenSystem, LevelGenSystem::(ano...
-#include "logger.h"             // for DEBUG
-#include "memory.hpp"           // for write_mem_prot, read_u64, read_u8
-#include "movable.hpp"          // for Movable
-#include "movable_behavior.hpp" // for init_behavior_hooks
-#include "render_api.hpp"       // for init_render_api_hooks
-#include "savedata.hpp"         // for SaveData
-#include "search.hpp"           // for get_address
-#include "spawn_api.hpp"        // for init_spawn_hooks
-#include "strings.hpp"          // for strings_init
-#include "thread_utils.hpp"     // for OnHeapPointer
-#include "virtual_table.hpp"    // for get_virtual_function_address, VTABLE...
+#include "entities_chars.hpp"    // for Player
+#include "entity.hpp"            // for to_id, Entity, HookWithId, EntityDB
+#include "entity_hooks_info.hpp" // for Player
+#include "game_manager.hpp"      // for get_game_manager, GameManager, SaveR...
+#include "items.hpp"             // for Items, SelectPlayerSlot
+#include "level_api.hpp"         // for LevelGenSystem, LevelGenSystem::(ano...
+#include "logger.h"              // for DEBUG
+#include "memory.hpp"            // for write_mem_prot, read_u64, read_u8
+#include "movable.hpp"           // for Movable
+#include "movable_behavior.hpp"  // for init_behavior_hooks
+#include "render_api.hpp"        // for init_render_api_hooks
+#include "savedata.hpp"          // for SaveData
+#include "search.hpp"            // for get_address
+#include "spawn_api.hpp"         // for init_spawn_hooks
+#include "strings.hpp"           // for strings_init
+#include "thread_utils.hpp"      // for OnHeapPointer
+#include "virtual_table.hpp"     // for get_virtual_function_address, VTABLE...
 
 uint16_t StateMemory::get_correct_ushabti() // returns animation_frame of ushabti
 {
